+++ chirho/dynamical/handlers/__init__.py
@@ -1,12 +1,12 @@
 from ..internals.backend import Solver  # noqa: F401
-from .dynamical import SimulatorEventLoop  # noqa: F401
+from .event_loop import SimulatorEventLoop  # noqa: F401
 from .interruption import (  # noqa: F401
     DynamicInterruption,
     DynamicIntervention,
     Interruption,
+    StaticBatchObservation,
     StaticInterruption,
     StaticIntervention,
     StaticObservation,
-    StaticBatchObservation,
 )
 from .trace import DynamicTrace  # noqa: F401
